move export files into one location in lib
